+++ src/app/models/book_view.py
@@ -4,7 +4,7 @@ from app.db.base_class import Base
 
 class BookView(Base):
     view_id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(BigInteger, ForeignKey("user.user_id"), nullable=True)
-    book_id = Column(BigInteger, ForeignKey("book.book_id"), nullable=False)
+    user_id = Column(Integer, ForeignKey("user.user_id"), nullable=True)
+    book_id = Column(Integer, ForeignKey("book.book_id"), nullable=False)
     
     viewed_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
